Removed Command Register Legacy Code
In a previous sprint/pr we merged in the working config based TYPE +
COMMAND system. The legacy code is causing major confusion when working
with the new system.

I testd by deleteing the command_registery table and wathced the logs as
I used the site and no logs.

I went to app and used ag to search the entire app file for
CommandRegistry to see if anything was invoking it or referencing it and
only the only files that referenced it for for itself.

This removes it and gets us closer to a clean system.